feat: handling for "`.`" -delimited column names
`extractNormalizedInsertQueryAndColumns` cleans column names but it misses a case where the column name is `.`. This is an issue for clickhouse nested columns. 

For example, go-gorm will produce sql with column names liek `events`.`attributes` and `links`.`span_id`. While the leading and trailing backticks are removed, the inner backticks remain which causes a `block cannot be sorted - missing columns in requested order` error.